mirror: API + structure polish (easy-to-use entry point, leaner modules)
- Rename the misleading _cli-suffixed programmatic entry points to clean verbs:
  solve_fixed_boundary / solve_free_boundary / solve_beta_scan (no CLI uses
  them; mirror was never released, so no back-compat aliases).
- Add the one-call convenience solve_fixed_boundary_from_radius(radius, config)
  so a fixed-boundary solve is one line instead of six setup calls.
- Merge exterior_bie.py into exterior.py (13 -> 12 modules) and drop the stale
  'virtual_casing_jax can be attached' aspirational note (never wired) and the
  redundant AxisymmetricExteriorVacuum alias.
- Route MirrorConvergenceError through the core VmecError taxonomy for uniform
  zero-crash handling.
- Promote the restart / beta-scan helpers the examples use into the public API.

All gates: ruff + mypy clean; tests/mirror green; grep confirms no _cli /
exterior_bie / alias / virtual_casing_jax remnants.
